requirements.txt
diffusers
transformers
accelerate
peft
bitsandbytes
pandas
gtts
pillow
gradio
opencv-python
controlnet_aux
huggingface-hub
numpy
cachetools
fsspec
safetensors
timm 
soundfile 
clip @ git+https://github.com/openai/CLIP.git
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git